Child names can be showed in ListView. Interface and GetName class removed
